% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_brick.R
\name{make_brick}
\alias{make_brick}
\title{Convenience function to make a RasterBrick}
\usage{
make_brick(x, ...)
}
\arguments{
\item{x}{A file path, or a URL to a file}

\item{...}{Further args passed on to \code{\link[raster]{brick}}}
}
\value{
An object of \code{\link[raster]{RasterBrick-class}}
}
\description{
Convenience function to make a RasterBrick
}
\examples{
\dontrun{
zip <- system.file("examples", "prismrain.zip", package = "timeraster")
dir <- paste0(tempdir(), "/prismrain")
dir <- "prismrain"
dir.create(dir)
unzip(zip, exdir = dir)
files <- list.files(dir, full.names = TRUE, pattern = ".tif$", all.files = TRUE)

# Create raster stack from file paths
res <- make_brick(files[1])

# native if TRUE uses raster package, if FALSE uses rgdal
make_brick(files[1], native = FALSE)
make_brick(files[1], native = TRUE)

# from a NetCDF file
## dates available
url <- "ftp://ftp.cdc.noaa.gov/Datasets/noaa.oisst.v2.highres/sst.day.mean.2016.v2.nc"
tf <- tempfile(fileext = ".nc")
download.file(url, tf, mode = 'wb')
x <- make_brick(tf)
time_raster(x)

nc <- system.file("examples", "prismrain.nc", package = "timeraster")
make_brick(nc)
}
}
\seealso{
\code{\link[raster]{brick}}
}

