% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_raster.R
\name{make_raster}
\alias{make_raster}
\title{Convenience function to make a RasterStack}
\usage{
make_raster(x, ...)
}
\arguments{
\item{x}{Vector or list of file paths}

\item{...}{Further args passed on to \code{\link[raster]{stack}}}
}
\value{
An object of \code{\link[raster]{RasterStack-class}}
}
\description{
Convenience function to make a RasterStack
}
\examples{
\dontrun{
zip <- system.file("examples", "prismrain.zip", package = "timeraster")
dir <- paste0(tempdir(), "/prismrain")
dir <- "prismrain"
dir.create(dir)
unzip(zip, exdir = dir)
files <- list.files(dir, full.names = TRUE, pattern = ".tif$", all.files = TRUE)

# Create raster stack from file paths
make_raster(files) \%>\% time_raster()
}
}
\seealso{
\code{\link[raster]{stack}}
}

