% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{TimeRaster-Class}
\alias{TimeRaster-Class}
\alias{[[,TimeRaster,ANY,ANY-method}
\alias{cellStats,TimeRaster-method}
\alias{getTS}
\alias{show,TimeRaster-method}
\title{An S4 class to represent A Time Raster object}
\usage{
\S4method{show}{TimeRaster}(object)

\S4method{[[}{TimeRaster,ANY,ANY}(x, i, j, ...)

getTS(object, ...)

\S4method{cellStats}{TimeRaster}(x, stat = "mean", na.rm = TRUE,
  asSample = TRUE, ...)
}
\arguments{
\item{object}{an object}

\item{x}{xxxx}

\item{i}{xxxx}

\item{j}{xxxx}

\item{...}{xxxx}

\item{stat}{xxxx}

\item{na.rm}{xxxx}

\item{asSample}{xxxx}
}
\description{
An S4 class to represent A Time Raster object
}
\section{Slots}{

\describe{
\item{\code{ts}}{An xts object}
}}
\keyword{methods}

