% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_brick.R
\name{make_brick}
\alias{make_brick}
\title{Convenience function to make a RasterBrick}
\usage{
make_brick(x, ...)
}
\arguments{
\item{x}{Vector or list of file paths}

\item{...}{Further args passed on to \code{\link[raster]{brick}}}
}
\value{
An object of \code{\link[raster]{RasterBrick-class}}
}
\description{
Convenience function to make a RasterBrick
}
\examples{
\dontrun{
zip <- system.file("examples", "prismrain.zip", package = "timeraster")
dir <- paste0(tempdir(), "/prismrain")
dir <- "prismrain"
dir.create(dir)
unzip(zip, exdir = dir)
files <- list.files(dir, full.names = TRUE, pattern = ".tif$", all.files = TRUE)

# Create raster stack from file paths
make_brick(files[1])
}
}
\seealso{
\code{\link[raster]{brick}}
}

