% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_raster.R
\name{time_raster}
\alias{time_raster}
\title{Time raster}
\usage{
time_raster(x, dates)
}
\arguments{
\item{x}{one of \code{\link[raster]{RasterStack-class}} or
\code{\link[raster]{RasterBrick-class}}}

\item{dates}{An object with dates, one of object of class
\code{\link[xts]{xts}}, or a \code{character} or \code{Date} class
vector of dates. If \code{character}, coerced internally to \code{Date}}
}
\value{
An object of class \code{\link{TimeRaster-class}}
}
\description{
Time raster
}
\examples{
\dontrun{
zip <- system.file("examples", "prismrain.zip", package = "timeraster")
dir <- paste0(tempdir(), "/prismrain")
dir <- "prismrain"
dir.create(dir)
unzip(zip, exdir = dir)
files <- list.files(dir, full.names = TRUE, pattern = ".tif$", all.files = TRUE)

# Create raster stack from file paths
rf <- raster::stack(files)

# Create a time series object
library("xts")
ts <- xts(1:365, as.Date("2014-01-01") + 0:364)

# Create TimeRaster object
## Using new()
rftr <- new("TimeRaster", rf, ts = ts)

## Use convenience function time_raster()
### With list of files as input - & xts object
rftr <- time_raster(files, ts)
### With RasterStack class as input - & xts object
rftr <- time_raster(rf, ts)
### With RasterStack class as input - & Date class object
rftr <- time_raster(rf, as.Date("2014-01-01") + 0:364)

# plot data
plot(rftr[["2014-10-01 TO 2014-10-03"]])
plot(rftr[["UPTO MONTHS"]])
}
}

